requirements.txt
fire==0.5.0
llama_cpp_python==0.2.32
matplotlib==3.8.2
pandas==2.1.4
pingouin==0.5.4
seaborn==0.13.0
statsmodels==0.14.1
tqdm==4.66.1